+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `scalr_workspace`: new attribute `terragrunt` ([#390](https://github.com/Scalr/terraform-provider-scalr/pull/390))
 - `scalr_environment` and `data.scalr_environment`: new attribute `mask_sensitive_output` ([#389](https://github.com/Scalr/terraform-provider-scalr/pull/389))
+- **New resource:** `scalr_integration_infracost` ([#CHANGE](https://github.com/Scalr/terraform-provider-scalr/pull/CHANGE)
+- **New data source:**  `scalr_integration_infracost` ([#CHANGE](https://github.com/Scalr/terraform-provider-scalr/pull/CHANGE))
 
 ### Changed